+++ x/evm/evmmodule/genesis_test.go
@@ -76,9 +76,9 @@ func (s *Suite) TestExportInitGenesis() {
 	s.Require().NoError(err)
 
 	// Send fungible token coins from bank to evm
-	_, err = deps.EvmKeeper.SendFunTokenToEvm(
+	_, err = deps.EvmKeeper.ConvertCoinToEvm(
 		deps.Ctx,
-		&evm.MsgSendFunTokenToEvm{
+		&evm.MsgConvertCoinToEvm{
 			Sender:    deps.Sender.NibiruAddr.String(),
 			BankCoin:  sdk.Coin{Denom: "unibi", Amount: math.NewInt(amountToSendC.Int64())},
 			ToEthAddr: eth.MustNewHexAddrFromStr(toUserC.String()),