+++ x/evm/keeper/erc20.go
@@ -55,10 +55,9 @@ func (e erc20Calls) Mint(
 ) (evmResp *evm.MsgEthereumTxResponse, err error) {
 	input, err := e.ABI.Pack("mint", to, amount)
 	if err != nil {
-		return
+		return nil, fmt.Errorf("failed to pack ABI args: %w", err)
 	}
-	commit := true
-	return e.CallContractWithInput(ctx, from, &contract, commit, input)
+	return e.CallContractWithInput(ctx, from, &contract, true, input)
 }
 
 /*
@@ -74,13 +73,23 @@ Transfer implements "ERC20.transfer"
 func (e erc20Calls) Transfer(
 	contract, from, to gethcommon.Address, amount *big.Int,
 	ctx sdk.Context,
-) (evmResp *evm.MsgEthereumTxResponse, err error) {
+) (out bool, err error) {
 	input, err := e.ABI.Pack("transfer", to, amount)
 	if err != nil {
-		return
+		return false, fmt.Errorf("failed to pack ABI args: %w", err)
 	}
-	commit := true
-	return e.CallContractWithInput(ctx, from, &contract, commit, input)
+	resp, err := e.CallContractWithInput(ctx, from, &contract, true, input)
+	if err != nil {
+		return false, err
+	}
+
+	var erc20Bool ERC20Bool
+	err = e.ABI.UnpackIntoInterface(&erc20Bool, "transfer", resp.Ret)
+	if err != nil {
+		return false, err
+	}
+
+	return erc20Bool.Value, nil
 }
 
 // BalanceOf retrieves the balance of an ERC20 token for a specific account.
@@ -339,31 +348,31 @@ func (k Keeper) loadERC20Uint8(
 
 func (k Keeper) LoadERC20BigInt(
 	ctx sdk.Context,
-	erc20Abi *gethabi.ABI,
-	erc20Contract gethcommon.Address,
+	abi *gethabi.ABI,
+	contract gethcommon.Address,
 	methodName string,
 	args ...any,
 ) (out *big.Int, err error) {
-	commit := false
 	res, err := k.CallContract(
 		ctx,
-		erc20Abi,
+		abi,
 		evm.EVM_MODULE_ADDRESS,
-		&erc20Contract,
-		commit,
+		&contract,
+		false,
 		methodName,
 		args...,
 	)
 	if err != nil {
-		return out, err
+		return nil, err
 	}
 
-	erc20Val := new(ERC20BigInt)
-	err = erc20Abi.UnpackIntoInterface(
-		erc20Val, methodName, res.Ret,
+	erc20BigInt := new(ERC20BigInt)
+	err = abi.UnpackIntoInterface(
+		erc20BigInt, methodName, res.Ret,
 	)
 	if err != nil {
-		return out, err
+		return nil, err
 	}
-	return erc20Val.Value, err
+
+	return erc20BigInt.Value, nil
 }