+++ x/evm/keeper/funtoken_from_coin.go
@@ -14,48 +14,50 @@ import (
 	"github.com/NibiruChain/nibiru/v2/x/evm/embeds"
 )
 
-func (k *Keeper) CreateFunTokenFromCoin(
+func (k *Keeper) createFunTokenFromCoin(
 	ctx sdk.Context, bankDenom string,
-) (funtoken evm.FunToken, err error) {
+) (funtoken *evm.FunToken, err error) {
 	// 1 | Coin already registered with FunToken?
 	if funtokens := k.FunTokens.Collect(ctx, k.FunTokens.Indexes.BankDenom.ExactMatch(ctx, bankDenom)); len(funtokens) > 0 {
-		return funtoken, fmt.Errorf("funtoken mapping already created for bank denom \"%s\"", bankDenom)
+		return nil, fmt.Errorf("funtoken mapping already created for bank denom \"%s\"", bankDenom)
 	}
 
 	// 2 | Check for denom metadata in bank state
-	bankCoin, isFound := k.bankKeeper.GetDenomMetaData(ctx, bankDenom)
+	bankMetadata, isFound := k.bankKeeper.GetDenomMetaData(ctx, bankDenom)
 	if !isFound {
-		return funtoken, fmt.Errorf("bank coin denom should have bank metadata for denom \"%s\"", bankDenom)
+		return nil, fmt.Errorf("bank coin denom should have bank metadata for denom \"%s\"", bankDenom)
 	}
 
 	// 3 | deploy ERC20 for metadata
-	erc20Addr, err := k.DeployERC20ForBankCoin(ctx, bankCoin)
+	erc20Addr, err := k.deployERC20ForBankCoin(ctx, bankMetadata)
 	if err != nil {
-		return funtoken, errors.Wrap(err, "failed to deploy ERC20 for bank coin")
+		return nil, errors.Wrap(err, "failed to deploy ERC20 for bank coin")
 	}
 
 	// 4 | ERC20 already registered with FunToken?
 	if funtokens := k.FunTokens.Collect(ctx, k.FunTokens.Indexes.ERC20Addr.ExactMatch(ctx, erc20Addr)); len(funtokens) > 0 {
-		return funtoken, fmt.Errorf("funtoken mapping already created for ERC20 \"%s\"", erc20Addr.Hex())
+		return nil, fmt.Errorf("funtoken mapping already created for ERC20 \"%s\"", erc20Addr.Hex())
 	}
 
 	// 5 | Officially create the funtoken mapping
-	funtoken = evm.FunToken{
+	funtoken = &evm.FunToken{
 		Erc20Addr:      eth.NewHexAddr(erc20Addr),
 		BankDenom:      bankDenom,
 		IsMadeFromCoin: true,
 	}
 
 	return funtoken, k.FunTokens.SafeInsert(
-		ctx, funtoken.Erc20Addr.ToAddr(),
+		ctx, erc20Addr,
 		funtoken.BankDenom,
 		funtoken.IsMadeFromCoin,
 	)
 }
 
-func (k *Keeper) DeployERC20ForBankCoin(
+func (k *Keeper) deployERC20ForBankCoin(
 	ctx sdk.Context, bankCoin bank.Metadata,
 ) (erc20Addr gethcommon.Address, err error) {
+	erc20Addr = crypto.CreateAddress(evm.EVM_MODULE_ADDRESS, k.GetAccNonce(ctx, evm.EVM_MODULE_ADDRESS))
+
 	// bank.Metadata validation guarantees that both "Base" and "Display" denoms
 	// pass "sdk.ValidateDenom" and that the "DenomUnits" slice has exponents in
 	// ascending order with at least one element, which must be the base
@@ -71,9 +73,6 @@ func (k *Keeper) DeployERC20ForBankCoin(
 	if err != nil {
 		return gethcommon.Address{}, errors.Wrap(err, "failed to pack ABI args")
 	}
-
-	erc20Addr = crypto.CreateAddress(evm.EVM_MODULE_ADDRESS, k.GetAccNonce(ctx, evm.EVM_MODULE_ADDRESS))
-
 	bytecodeForCall := append(embeds.SmartContract_ERC20Minter.Bytecode, packedArgs...)
 
 	// nil address for contract creation